feat_7a: add user roles check

+++ src/articles/articles.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { Article, Tag } from '@prisma/client';
 
 import * as C from '../constants';
@@ -10,7 +10,7 @@ import { UpdateArticleDto } from './dto/update-article.dto';
 
 @Injectable()
 export class ArticlesService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService) { }
 
   private safeArticle({
     createdAt,
@@ -108,11 +108,11 @@ export class ArticlesService {
         updatedAt: now,
         tags: dto.tags
           ? {
-              connectOrCreate: dto.tags.map((name) => ({
-                where: { name },
-                create: { name },
-              })),
-            }
+            connectOrCreate: dto.tags.map((name) => ({
+              where: { name },
+              create: { name },
+            })),
+          }
           : undefined,
       },
       include: { tags: true },
@@ -120,9 +120,13 @@ export class ArticlesService {
     return this.safeArticle(article);
   }
 
-  async update(id: string, dto: UpdateArticleDto): Promise<T.ArticleType> {
+  async update(id: string, dto: UpdateArticleDto, user: T.TokenPayloadType): Promise<T.ArticleType> {
     const exists = await this.prisma.article.findUnique({ where: { id } });
     if (!exists) throw new NotFoundException(C.ARTICLE_NOT_FOUND);
+
+    if (user.role === C.EDITOR && exists.authorId !== user.userId) {
+      throw new ForbiddenException(C.EDIT_EXCEPTION);
+    }
 
     const updated = await this.prisma.article.update({
       where: { id },
@@ -131,12 +135,12 @@ export class ArticlesService {
         updatedAt: new Date(),
         tags: dto.tags
           ? {
-              set: [],
-              connectOrCreate: dto.tags.map((name) => ({
-                where: { name },
-                create: { name },
-              })),
-            }
+            set: [],
+            connectOrCreate: dto.tags.map((name) => ({
+              where: { name },
+              create: { name },
+            })),
+          }
           : undefined,
       },
       include: { tags: true },
@@ -145,12 +149,15 @@ export class ArticlesService {
     return this.safeArticle(updated);
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: T.TokenPayloadType) {
     const exists = await this.prisma.article.findUnique({ where: { id } });
     if (!exists) throw new NotFoundException(C.ARTICLE_NOT_FOUND);
 
-    await this.prisma.article.delete({ where: { id } });
+    if (user.role !== C.ADMIN && exists.authorId !== user.userId) {
+      throw new ForbiddenException(C.ARTICLE_DELETE_EXCEPTION);
+    }
 
+    await this.prisma.article.delete({ where: { id } });
     return null;
   }
 }
